[ls-build-html] refact: move copySyncHelper to own file, rm siteBuilder fn
